Commit
Versão web com o link da Amazon funcionando.
Backend já conectado.

+++ backend/database/migrations/2021_09_03_000003_create_titulos_table.php
@@ -9,8 +9,8 @@ class CreateTitulosTable extends Migration {
         Schema::create('titulos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->integer('editora_id')->unsigned();
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('editora_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->string('observacao')->nullable();
             $table->timestamps();
             $table->softDeletes();
