+++ backend/database/migrations/2021_09_03_000004_create_volumes_table.php
@@ -10,10 +10,10 @@ class CreateVolumesTable extends Migration {
             $table->id();
             $table->string('nome');
             $table->double('preco', 8, 2);
-            $table->integer('titulo_id')->unsigned();
+            $table->bigInteger('titulo_id')->unsigned();
             $table->string('capa')->nullable();
             $table->date('data_lancamento')->nullable();
-            $table->integer('ordem');
+            $table->bigInteger('ordem');
             $table->string('observacao')->nullable();
             $table->timestamps();
             $table->softDeletes();
