+++ backend/database/migrations/2021_09_03_000011_create_usuario_titulos_favoritos_table.php
@@ -8,8 +8,8 @@ class CreateUsuarioTitulosFavoritosTable extends Migration {
     public function up() {
         Schema::create('usuario_titulos_favoritos', function (Blueprint $table) {
             $table->id();
-            $table->integer('usuario_id')->unsigned();
-            $table->integer('titulo_id')->unsigned();
+            $table->bigInteger('usuario_id')->unsigned();
+            $table->bigInteger('titulo_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
