+++ backend/database/migrations/2021_09_03_000012_create_usuario_volumes_favoritos_table.php
@@ -8,8 +8,8 @@ class CreateUsuarioVolumesFavoritosTable extends Migration {
     public function up() {
         Schema::create('usuario_volumes_favoritos', function (Blueprint $table) {
             $table->id();
-            $table->integer('usuario_id')->unsigned();
-            $table->integer('volume_id')->unsigned();
+            $table->bigInteger('usuario_id')->unsigned();
+            $table->bigInteger('volume_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
